pyproject.toml
[project]
name = "unity"
version = "0.1.0"
description = "A Discord bot made for fun."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.1",
]

[project.scripts]
test = "test:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
